tests

+++ src/party/party.gateway.ts
@@ -15,22 +15,25 @@ export class PartyGateway {
 
   @SubscribeMessage('create_party')
   create(@ConnectedSocket() client: Socket) {
-    this.logger.debug('create_party');
     const email = client.handshake.auth.email;
+    if (!email) return false;
+    this.logger.debug('create_party');
     return this.partyService.create({ email });
   }
 
   @SubscribeMessage('get_party')
   findOne(@ConnectedSocket() client: Socket) {
-    this.logger.debug('get_party');
     const email = client.handshake.auth.email;
+    if (!email) return false;
+    this.logger.debug('get_party');
     return this.partyService.findOne({ email });
   }
 
   @SubscribeMessage('remove_party')
   remove(@ConnectedSocket() client: Socket) {
-    this.logger.debug('remove_party');
     const email = client.handshake.auth.email;
+    if (!email) return false;
+    this.logger.debug('remove_party');
     return this.partyService.remove({ email });
   }
 
@@ -39,29 +42,33 @@ export class PartyGateway {
     @MessageBody() invitedEmail: string,
     @ConnectedSocket() client: Socket,
   ) {
-    this.logger.debug('invite_to_party');
     const email = client.handshake.auth.email;
+    if (!email) return false;
+    this.logger.debug('invite_to_party');
     return this.partyService.invite({ leaderEmail: email, invitedEmail });
   }
 
   @SubscribeMessage('kick_from_party')
   kick(@MessageBody() kickedEmail: string, @ConnectedSocket() client: Socket) {
-    this.logger.debug('kick_from_party');
     const email = client.handshake.auth.email;
+    if (!email) return false;
+    this.logger.debug('kick_from_party');
     return this.partyService.kick({ leaderEmail: email, kickedEmail });
   }
 
   @SubscribeMessage('quit_party')
   quit(@MessageBody() partyId: number, @ConnectedSocket() client: Socket) {
-    this.logger.debug('quit_party');
     const email = client.handshake.auth.email;
+    if (!email) return false;
+    this.logger.debug('quit_party');
     return this.partyService.quitParty({ email: email, partyId });
   }
 
   @SubscribeMessage('join_party')
   join(@MessageBody() partyId: number, @ConnectedSocket() client: Socket) {
-    this.logger.debug('join_party');
     const email = client.handshake.auth.email;
+    if (!email) return false;
+    this.logger.debug('join_party');
     return this.partyService.joinParty({ email, partyId });
   }
 }
